OC-256 Refactors code for 1 webserviceTemplateFactory
