+++ cucumber-tests-platform/src/test/java/com/alliander/osgp/platform/cucumber/config/ws/tariffswitching/TariffSwitchingAdhocManagementWebServiceConfig.java
@@ -14,7 +14,7 @@ import org.springframework.oxm.jaxb.Jaxb2Marshaller;
 import org.springframework.ws.server.endpoint.adapter.method.MarshallingPayloadMethodProcessor;
 
 import com.alliander.osgp.platform.cucumber.config.ws.BaseWebServiceConfig;
-import com.alliander.osgp.platform.cucumber.support.ws.WebServiceTemplateFactory;
+import com.alliander.osgp.shared.infra.ws.WebServiceTemplateFactory;
 
 @Configuration
 public class TariffSwitchingAdhocManagementWebServiceConfig extends BaseWebServiceConfig {
@@ -27,13 +27,9 @@ public class TariffSwitchingAdhocManagementWebServiceConfig extends BaseWebServi
 
     @Bean
     public WebServiceTemplateFactory tariffSwitchingAdHocManagementWstf() {
-        return new WebServiceTemplateFactory.Builder().setMarshaller(this.tariffSwitchingAdHocManagementMarshaller())
-                .setMessageFactory(this.messageFactory())
-                .setDefaultUri(this.baseUri.concat(this.webserviceTemplateDefaultUriTariffSwitchingAdHocManagement))
-                .setKeyStoreType(this.webserviceKeystoreType).setKeyStoreLocation(this.webserviceKeystoreLocation)
-                .setKeyStorePassword(this.webserviceKeystorePassword)
-                .setTrustStoreFactory(this.webServiceTrustStoreFactory()).setApplicationName(this.applicationName)
-                .build();
+        return new WebServiceTemplateFactory(this.tariffSwitchingAdHocManagementMarshaller(), this.messageFactory(),
+                this.baseUri.concat(this.webserviceTemplateDefaultUriTariffSwitchingAdHocManagement), this.webserviceKeystoreType,
+                this.webserviceKeystoreLocation, this.webserviceKeystorePassword, this.webServiceTrustStoreFactory(), this.applicationName);
     }
 
     /**
