+++ cucumber-tests-platform/src/test/java/com/alliander/osgp/platform/cucumber/config/ws/tariffswitching/TariffSwitchingScheduleManagementWebServiceConfig.java
@@ -14,7 +14,7 @@ import org.springframework.oxm.jaxb.Jaxb2Marshaller;
 import org.springframework.ws.server.endpoint.adapter.method.MarshallingPayloadMethodProcessor;
 
 import com.alliander.osgp.platform.cucumber.config.ws.BaseWebServiceConfig;
-import com.alliander.osgp.platform.cucumber.support.ws.WebServiceTemplateFactory;
+import com.alliander.osgp.shared.infra.ws.WebServiceTemplateFactory;
 
 @Configuration
 public class TariffSwitchingScheduleManagementWebServiceConfig extends BaseWebServiceConfig {
@@ -27,13 +27,9 @@ public class TariffSwitchingScheduleManagementWebServiceConfig extends BaseWebSe
 
     @Bean
     public WebServiceTemplateFactory tariffSwitchingScheduleManagementWstf() {
-        return new WebServiceTemplateFactory.Builder().setMarshaller(this.tariffSwitchingScheduleManagementMarshaller())
-                .setMessageFactory(this.messageFactory())
-                .setDefaultUri(this.baseUri.concat(this.webserviceTemplateDefaultUriTariffSwitchingScheduleManagement))
-                .setKeyStoreType(this.webserviceKeystoreType).setKeyStoreLocation(this.webserviceKeystoreLocation)
-                .setKeyStorePassword(this.webserviceKeystorePassword)
-                .setTrustStoreFactory(this.webServiceTrustStoreFactory()).setApplicationName(this.applicationName)
-                .build();
+        return new WebServiceTemplateFactory(this.tariffSwitchingScheduleManagementMarshaller(), this.messageFactory(),
+                this.baseUri.concat(this.webserviceTemplateDefaultUriTariffSwitchingScheduleManagement), this.webserviceKeystoreType,
+                this.webserviceKeystoreLocation, this.webserviceKeystorePassword, this.webServiceTrustStoreFactory(), this.applicationName);
     }
 
     /**
