+++ cucumber-tests-platform/src/test/java/com/alliander/osgp/platform/cucumber/support/ws/tariffswitching/TariffSwitchingAdHocManagementClient.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.alliander.osgp.platform.cucumber.support.ws.BaseClient;
-import com.alliander.osgp.platform.cucumber.support.ws.WebServiceTemplateFactory;
+import com.alliander.osgp.shared.infra.ws.WebServiceTemplateFactory;
 
 @Component
 public class TariffSwitchingAdHocManagementClient extends BaseClient {
