+++ cucumber-tests-platform/src/test/java/com/alliander/osgp/platform/cucumber/support/ws/tariffswitching/TariffSwitchingScheduleManagementClient.java
@@ -19,8 +19,8 @@ import com.alliander.osgp.adapter.ws.schema.tariffswitching.schedulemanagement.S
 import com.alliander.osgp.adapter.ws.schema.tariffswitching.schedulemanagement.SetScheduleRequest;
 import com.alliander.osgp.adapter.ws.schema.tariffswitching.schedulemanagement.SetScheduleResponse;
 import com.alliander.osgp.platform.cucumber.support.ws.BaseClient;
-import com.alliander.osgp.platform.cucumber.support.ws.WebServiceSecurityException;
-import com.alliander.osgp.platform.cucumber.support.ws.WebServiceTemplateFactory;
+import com.alliander.osgp.shared.exceptionhandling.WebServiceSecurityException;
+import com.alliander.osgp.shared.infra.ws.WebServiceTemplateFactory;
 
 @Component
 public class TariffSwitchingScheduleManagementClient extends BaseClient {
